python/requirements.txt
inference
python-dotenv